Logical ops (#58)
This PR adds `AND`, `OR`, and `NOT` opcodes.

These operations are designed for basically one purpose: making
conditionals of the form `(cond && a) || (!cond && b)`. As such, they
have the following semantics:

- `a && b ⇒ if a == 0.0 { a } else { b }`
- `a || b ⇒ if a != 0.0 { a } else { b }`

The operators are also designed to work well with tape simplification
(through the `TracingEvaluator`): if `cond` is unambiguously `0.0` or
unambiguously non-zero, then the conditional statement can be collapsed
into either `a` or `b`.