Add `vars` to `Function`; split `VmFunction/Tape` (#197)
We want to be able to get `vars` from a `Function` without building a
tape. However, upon adding it to `trait Function`, there's ambiguity
about whether to call `Function::vars` or `Tape::vars` on a
`GenericVmFunction`, which implements both traits. To fix this
ambiguity, we add a new `GenericVmTape` type.