Use Rayon on the web (#207)
Using
[`wasm-bindgen-rayon`](https://github.com/RReverser/wasm-bindgen-rayon/),
we can render with the global Rayon thread pool (instead of hand-rolling
a tile-based renderer). This is significantly faster!

Unfortunately, as part of the Web Ecosystem™, everything is janky and
fragile. The code includes workarounds for the following issues:

- https://github.com/rustwasm/wasm-pack/issues/1402 (`wasm-pack` changed
things, breaking Webpack)
- https://github.com/rustwasm/wasm-bindgen/issues/4233 (`wasm-bindgen`
fixed what they believe to be a bug, breaking `wasm-bindgen-rayon` in
the process)