Fix NAN propagation through interval and / or opcodes (#245)
`[NaN, NaN]` does not mean "this interval is guaranteed to be NaN";
instead it means "this interval may have any value". As such, we can't
short-circuit when doing interval evaluation of `And` and `Or` opcodes.

A `NaN` in either argument must be propagated, and we must set
the `Choice::Both` flag.

Fixes #240 (for real this time)