Make cli demo compatible with wasm32-wasi
While the wasm32-unknown-unknown environment makes little sense due to
the absence of any file handling, there is a case for supporting wasm32
anyways. With the wasi system interface the underlying platform provides
all the IO required, making this a viable `std` target. In particular,
WebAssembly ensures a high degree of reproducibility (especially under
implementations that ensure canonical NaN handling).

As a demo, we can pack the compiled module into an HTML page and
polyfill the file data in Javascript such as bjorn3's wasi shim.

I've only locally verified against Rust's `wasm32-wasip1` target but
forsee no reason for newer wasi preview versions to work any worse when
they are stabilized. To indicate compatibility with future versions it
would be helpful to ensure CI coverage for this target, if maintenance
effort allows it.

The main overheads of such compatibility in the core of `fidget` and its
rhai embedding appear with the `wasm32` target architecture itself, not
the `-wasi` platform. This is already covered as a Tier 1 target. There
are some subtleties with libraries that the CLI may depend on in the
future. The common terminal API libraries that would be used for
interactive modes generally do not work with `wasm` targets, many depend
on UNIX-ly libc functionality etc. Again though, this would hold even
less true for `wasm32-unknown-unknown` where definitionally no system
interface can be defined at all.