Rewrite meshing (#281)
Previously, meshing used a homebrew threadpool to share fine-grained
work between threads. This is both scary (I don't trust the lock-free
code) and incompatible with WebAssembly (which stubs out `std::thread`).

This PR switches to using Rayon, using a map-reduce-ish strategy:

- We recurse down the octree _without doing evaluation_ (so it's cheap)
  to build a sufficiently large set of tasks
- These tasks are executed in parallel using Rayon
- Then, we merge the resulting octrees into the root (which is
  relatively cheap)

There are many other changes along the way, because this lets us
dramatically simplify the logic of recursion!

In addition, many of the meshing types are now parameterized with a
`const D: usize` to support 2D contouring in the future (see #223).

This is a noticeable speedup, testing
```
cargo run -pfidget-cli --release -- mesh -i models/cabin.rhai --scale 0.05 --center=0,0,-14 --depth 8 -N10
```

On `main` (1099d8f5), I see
```
[2025-03-21T21:27:37Z INFO  fidget_cli] Rendered 10x at 343.111 ms/iter
[2025-03-21T21:27:37Z INFO  fidget_cli]   Octree construction: 306.6612 ms/iter
[2025-03-21T21:27:37Z INFO  fidget_cli]   Mesh construction: 36.1389 ms/iter
```

In this branch, I see
```
[2025-03-21T21:27:11Z INFO  fidget_cli] Rendered 10x at 283.3219 ms/iter
[2025-03-21T21:27:11Z INFO  fidget_cli]   Octree construction: 227.1011 ms/iter
[2025-03-21T21:27:11Z INFO  fidget_cli]   Mesh construction: 55.8783 ms/iter
```

Note that mesh construction is **not parallelized yet**, so that
slowdown is expected. Even so, the faster octree construction more that
compensates!